% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroyLocalCopy.R
\name{destroyLocalCopy}
\alias{destroyLocalCopy}
\alias{destroyLocalCopy.default}
\alias{destroyLocalCopy.character}
\title{Delete local project tables}
\usage{
destroyLocalCopy(localdb, rcon)

\method{destroyLocalCopy}{default}(localdb, rcon)

\method{destroyLocalCopy}{character}(localdb, rcon)
}
\arguments{
\item{localdb}{\code{link[DBI]{dbConnect}} object, or perhaps local folder}

\item{rcon}{\code{link[redcapAPI]{redcapConnection}} object}
}
\value{
TRUE
}
\description{
Delete tables associated with locally stored REDCap project.
}
