% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{projectId}
\alias{projectId}
\title{Retrieve REDCap project_id}
\usage{
projectId(rcon)
}
\arguments{
\item{rcon}{\code{link[redcapAPI]{redcapConnection}} object}
}
\value{
character, REDCap project_id
}
\description{
Look up a \code{link[redcapAPI]{redcapConnection}} object's project_id.
}
