% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectTables.R
\name{projectTables}
\alias{projectTables}
\alias{projectTables.default}
\alias{projectTables.character}
\title{List local project tables}
\usage{
projectTables(localdb, rcon = NULL)

\method{projectTables}{default}(localdb, rcon = NULL)

\method{projectTables}{character}(localdb, rcon = NULL)
}
\arguments{
\item{localdb}{\code{link[DBI]{dbConnect}} object, or perhaps local folder}

\item{rcon}{\code{link[redcapAPI]{redcapConnection}} object}
}
\value{
character, table names
}
\description{
List tables associated with locally stored REDCap project.
}
