% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/READCap-package.R
\docType{package}
\name{READCap-package}
\alias{READCap}
\alias{READCap-package}
\title{Keep REDCap Synced Locally}
\description{
Use redcapAPI to download data from REDCap and sync it with a database backend.
}
\author{
\strong{Maintainer}: Cole Beck \email{cole.beck@vumc.org} (\href{https://orcid.org/0000-0002-6849-6255}{ORCID})

}
