% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLogs.R
\name{readLogs}
\alias{readLogs}
\alias{readLogs.default}
\alias{readLogs.character}
\title{Read Local REDCap Event Logs}
\usage{
readLogs(localdb, rcon)

\method{readLogs}{default}(localdb, rcon)

\method{readLogs}{character}(localdb, rcon)
}
\arguments{
\item{localdb}{\code{link[DBI]{dbConnect}} object, or perhaps local folder}

\item{rcon}{\code{link[redcapAPI]{redcapConnection}} object}
}
\value{
data.frame, event logs
}
\description{
List REDCap event logs.
}
