% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncLocalCopy.R
\name{syncLocalCopy}
\alias{syncLocalCopy}
\alias{syncLocalCopy.default}
\alias{syncLocalCopy.character}
\title{Sync local project tables}
\usage{
syncLocalCopy(localdb, rcon)

\method{syncLocalCopy}{default}(localdb, rcon)

\method{syncLocalCopy}{character}(localdb, rcon)
}
\arguments{
\item{localdb}{\code{link[DBI]{dbConnect}} object, or perhaps local folder}

\item{rcon}{\code{link[redcapAPI]{redcapConnection}} object}
}
\value{
numeric, number of records updated
}
\description{
Sync tables associated with locally stored REDCap project.
}
